Proper Implementation using Factory and Interface
This is a WIP, you will have to review it and make any necessary tweaks
to ensure it works properly.

Yes, I’ve removed the ability to pass the $title into the constructor.
That’s because you would have to put this logic into the $factory
itself (to take the second argument and attach it as a title property
of the object) and this just is not something the factory should be
responsible for.

I also relocated the factory instance to where you seem to bootstrap
everything, hopefully it’s being executed BEFORE the subclasses or you
may get an error.